backend/requirements.txt
Django==3.2.15
django-extensions==3.2.1
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
environ==1.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.8.5
pyflakes==2.5.0
PyJWT==2.5.0
pytz==2022.2.1
webcolors==1.12
